configure user routers & a[i end points

+++ app/routers/user.py
@@ -1,7 +1,7 @@
 from fastapi import Depends, HTTPException , APIRouter
 from sqlalchemy.orm import Session
 from app.db.dependancies import get_db
-from app.schemas.user import UserModel , UserRegisterModel , FullUserModel
+from app.schemas.user import LoginUserModel , RegisterUserModel , FullUserModel
 from app.services import user
 
 router = APIRouter(
@@ -13,18 +13,18 @@ router = APIRouter(
 
 
 @router.post("/add/",response_model=FullUserModel)
-def add_user(user_data:UserRegisterModel , db: Session = Depends(get_db)) -> FullUserModel:
+def add_user(user_data:RegisterUserModel , db: Session = Depends(get_db)) -> FullUserModel:
     """
     add new user
     """
-    user = user.get_user_by_username(db , user_data.username.lower())
-    if user:
+    user_exits = user.get_user_by_username(db , user_data.username.lower())
+    if user_exits:
         raise HTTPException(status_code=400, detail="Username already registered")
     elif user_data.password != user_data.confirm:
         raise HTTPException(status_code=400 , detail="parsed password & confirmed password don't match ")
 
     hashed_password = user.get_hashed_password(user_data.password).decode("utf-8")
-    user_data = UserModel(
+    user_data = LoginUserModel(
             username = user_data.username.lower() , 
             password = hashed_password
     )
@@ -49,10 +49,10 @@ def get_user_by_id(user_id:int , db:Session = Depends(get_db)) -> FullUserModel:
     """
     by id get user
     """
-    user = user.get_user_by_id(db , user_id = user_id)
-    if not user:
+    user_by_id = user.get_user_by_id(db , user_id = user_id)
+    if not user_by_id:
         raise HTTPException(status_code=400, detail=f"no user with id {user_id}")
-    return user
+    return user_by_id
 
 
 
@@ -63,6 +63,9 @@ def delete_user(user_id:int, db:Session = Depends(get_db)):
     """
     delete user with id
     """
+    user_by_id = user.get_user_by_id(db , user_id = user_id)
+    if not user_by_id:
+        raise HTTPException(status_code=400, detail=f"no user with id {user_id}")
     user.delete_user(db , user_id = user_id)
     return {"message":f"user with id {user_id} deleted"}
